# backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: Descubre República Dominicana — API
  version: 1.0.0
  description: API REST del portal turístico Descubre RD. Versionada en /api/v1. Autenticación via Bearer JWT (Supabase Auth).
  contact: 
    name: Equipo Técnico Descubre RD
    email: "tech@descubrerd.do"
  license: 
    name: Privativo — uso interno
servers:
  - url: "https://api.descubrerd.do/api/v1"
    description: Producción
  - url: "https://staging-api.descubrerd.do/api/v1"
    description: Staging
  - url: "http://localhost:3000/api/v1"
    description: Desarrollo local
...
components:
  securitySchemes: 
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT emitido por Supabase Auth. Header: Authorization: Bearer <token>"
...
paths:
...
  /sponsorship/serve/{slot_id}: 
    get: 
      tags: 
        - patrocinio
      summary: Entrega creatividades activas para un slot
      parameters: 
        - name: slot_id
          in: path
          required: true
          schema: 
            type: string
        - name: category
          in: query
          schema: 
            type: string
        - name: destination
          in: query
          schema: 
            type: string
        - name: limit
          in: query
          schema: 
            type: integer
            maximum: 10
            default: 1
      responses: 
        200: 
          description: Creatividades entregadas
  /sponsorship/telemetry: 
    post: 
      tags: 
        - patrocinio
      summary: "Registra telemetria de anuncio (impresion, clic, conversion)"
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - creative_id
                - slot_id
                - event_type
              properties: 
                creative_id: 
                  type: string
                  format: uuid
                slot_id: 
                  type: string
                event_type: 
                  type: string
                  enum: 
                    - impression
                    - click
                    - conversion
                session_id: 
                  type: string
                page: 
                  type: string
      responses: 
        200: 
          description: Telemetria registrada
...
  /sponsorship/campaigns: 
    post: 
      tags: 
        - patrocinio
      summary: Crea una campana publicitaria
      security: 
        - bearerAuth: []
      responses: 
        201: 
          description: Campana creada
...